requirements.txt
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3 